initial commit
Contains a working very basic blocking Discord client, but the
SNI client does not work quite yet.

The .proto file is read to generate gRPC code, but up until now
it is unclear how to bootstrap that code to tap into a running
SNI (Super Nintendo Interface) instance connected to a rom in
RetroArch (which works fine for the Dunka tracker).